refactor: move adapters into internal/adapter
## Summary

Move the adapter implementations into an explicit adapter package structure.

## Why

This is another step toward making the hexagonal structure visible in the codebase.

The controller, Kubernetes integration, and NATS integration are all adapters around the core application logic, so they fit better under `internal/adapter/inbound` and `internal/adapter/outbound` than as top-level internal packages.

## What Changed

- moved the controller code to `internal/adapter/inbound/controller`
- moved the Kubernetes adapter code to `internal/adapter/outbound/k8s`
- moved the NATS adapter code to `internal/adapter/outbound/nats`
- updated imports and related test setup to match the new package locations

## Follow-up TODOs

This commit makes the package layout clearer, but it also highlights boundaries that still need to be enforced:

- core must not depend on adapter code
- controller must not depend on other adapter code

Those concerns are intentionally left as follow-up work and are noted in the code for later cleanup.

## Notes

No intended behavior change. This is a structural refactor to align the package layout with the target architecture.
