+++ cmd/main.go
@@ -25,8 +25,6 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/WirelessCar/nauth/internal/domain"
-	"github.com/WirelessCar/nauth/internal/nats"
 	"sigs.k8s.io/controller-runtime/pkg/cache"
 
 	// Import all Kubernetes client auth plugins (e.g. Azure, GCP, OIDC, etc.)
@@ -45,11 +43,13 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/webhook"
 
 	"github.com/WirelessCar/nauth/api/v1alpha1"
-	"github.com/WirelessCar/nauth/internal/controller"
+	"github.com/WirelessCar/nauth/internal/adapter/inbound/controller"
+	"github.com/WirelessCar/nauth/internal/adapter/outbound/k8s"
+	"github.com/WirelessCar/nauth/internal/adapter/outbound/k8s/configmap"
+	"github.com/WirelessCar/nauth/internal/adapter/outbound/k8s/secret"
+	"github.com/WirelessCar/nauth/internal/adapter/outbound/nats"
 	"github.com/WirelessCar/nauth/internal/core"
-	"github.com/WirelessCar/nauth/internal/k8s"
-	"github.com/WirelessCar/nauth/internal/k8s/configmap"
-	"github.com/WirelessCar/nauth/internal/k8s/secret"
+	"github.com/WirelessCar/nauth/internal/domain"
 	// +kubebuilder:scaffold:imports
 )
