+++ internal/adapter/inbound/controller/account.go
@@ -21,8 +21,8 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/WirelessCar/nauth/internal/adapter/outbound/k8s" // TODO: Controller must not depend on other adapter code
 	"github.com/WirelessCar/nauth/internal/domain"
-	"github.com/WirelessCar/nauth/internal/k8s"
 	"github.com/WirelessCar/nauth/internal/ports/inbound"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
