+++ internal/adapter/inbound/controller/account_test.go
@@ -20,8 +20,8 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/WirelessCar/nauth/internal/adapter/outbound/k8s" // TODO: Controller must not depend on other adapter code
 	"github.com/WirelessCar/nauth/internal/domain"
-	"github.com/WirelessCar/nauth/internal/k8s"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"github.com/stretchr/testify/mock"
