+++ internal/adapter/outbound/k8s/errors.go
