+++ internal/adapter/outbound/nats/client.go
