+++ internal/core/secret.go
@@ -9,8 +9,8 @@ import (
 	"io"
 	"sync"
 
+	"github.com/WirelessCar/nauth/internal/adapter/outbound/k8s" // TODO: Core must not depend on adapter code
 	"github.com/WirelessCar/nauth/internal/domain"
-	"github.com/WirelessCar/nauth/internal/k8s"
 	"github.com/WirelessCar/nauth/internal/ports/outbound"
 	"github.com/nats-io/nkeys"
 	v1 "k8s.io/api/core/v1"
